Add files via upload
This is our final course project for the Python course "Analytics Programming" where we working on COVID-19 data by the NGO COVID ActNow of USA, we analyzed the data of all the states and find out information about deaths, cases, and information about hospitalizations. We also did sentiment analysis on Twitter data using Tweepy and analyzed what were the sentiments of people during covid 19 and when the vaccine was launched.

We saw people had positive sentiments after the roll-out of vaccine.